Rename tools submodule to utils to allow for development of tools plugin framework
