+++ dtocean_core/utils/electrical.py
