+++ example_data/fixed_tidal_fixed_layout_scenario.py
@@ -13,8 +13,8 @@ import pandas as pd
 from scipy.stats import multivariate_normal, norm
 from shapely.geometry import Point
 
-from dtocean_core.tools.moorings import get_moorings_tables
-from dtocean_core.tools.reliability import (get_reliability_tables,
+from dtocean_core.utils.moorings import get_moorings_tables
+from dtocean_core.utils.reliability import (get_reliability_tables,
                                             compdict_from_mock)
 
 # Note that the electrical folder in the test_data directory should be
@@ -1424,7 +1424,7 @@ test_data = {
              
 if __name__ == "__main__":
     
-    from dtocean_core.tools.files import pickle_test_data
+    from dtocean_core.utils.files import pickle_test_data
 
     file_path = os.path.abspath(__file__)
     pkl_path = pickle_test_data(file_path, test_data)
