+++ example_data/fixed_wave_staggered_layout_scenario.py
@@ -11,7 +11,7 @@ import pandas as pd
 
 from datetime import datetime, timedelta
 
-from dtocean_core.tools.moorings import get_moorings_tables
+from dtocean_core.utils.moorings import get_moorings_tables
 
 # Note that the electrical folder in the test_data directory should be
 # placed in the same folder as this file
@@ -849,7 +849,7 @@ test_data = {
              
 if __name__ == "__main__":
     
-    from dtocean_core.tools.files import pickle_test_data
+    from dtocean_core.utils.files import pickle_test_data
 
     file_path = os.path.abspath(__file__)
     pkl_path = pickle_test_data(file_path, test_data)
