+++ test_data/inputs_reliability.py
@@ -10,7 +10,7 @@ from collections import Counter
 
 import pandas as pd
 
-from dtocean_core.tools.reliability import (get_reliability_tables,
+from dtocean_core.utils.reliability import (get_reliability_tables,
                                             compdict_from_mock)
 
 this_dir = os.path.dirname(os.path.realpath(__file__))
@@ -85,7 +85,7 @@ test_data = {"device.system_type" : device_type,
 
 if __name__ == "__main__":
     
-    from dtocean_core.tools.files import pickle_test_data
+    from dtocean_core.utils.files import pickle_test_data
 
     file_path = os.path.abspath(__file__)
     pkl_path = pickle_test_data(file_path, test_data)
