+++ test_data/inputs_wp6.py
@@ -12,7 +12,7 @@ from collections import Counter
 import numpy as np
 import pandas as pd
 
-from dtocean_core.tools.reliability import (get_reliability_tables,
+from dtocean_core.utils.reliability import (get_reliability_tables,
                                             compdict_from_mock)
 
 from inputs_wp5 import (equipment_cable_burial,
@@ -869,7 +869,7 @@ test_data = {
              
 if __name__ == "__main__":
     
-    from dtocean_core.tools.files import pickle_test_data
+    from dtocean_core.utils.files import pickle_test_data
 
     file_path = os.path.abspath(__file__)
     pkl_path = pickle_test_data(file_path, test_data)
