+++ test_data/installation/inputs_wp5_electrical_star_fixed.py
@@ -537,7 +537,7 @@ test_data = {"component.rov" : equipment_rov,
 
 if __name__ == "__main__":
 
-    from dtocean_core.tools.files import pickle_test_data
+    from dtocean_core.utils.files import pickle_test_data
 
     file_path = os.path.abspath(__file__)
     pkl_path = pickle_test_data(file_path, test_data)
